refactor(stores,nats): reorganize into subpackages to satisfy folder-size gate (#1595)
Split src/lyra/infrastructure/stores into migrations/ + base/ subpackages:
- migrations/: agent_store_migrations.py, bot_store_migrations.py
- base/: sqlite_base.py, message_index.py, bot_agent_map.py

Split src/lyra/nats into audio/, stt/, image/ subpackages:
- audio/: nats_tts_client.py, nats_tts_codec.py, tts_engine_selector.py, tts_text_normalization.py
- stt/: nats_stt_client.py, nats_stt_codec.py, stt_helpers.py
- image/: nats_image_client.py, nats_image_codec.py

Update all internal imports and test imports. Remove folder exemptions.
