+++ src/lyra/infrastructure/stores/__init__.py
@@ -3,10 +3,11 @@ ADR-048 (absorbed into ADR-059)."""
 
 from lyra.infrastructure.stores.agent_store import AgentStore
 from lyra.infrastructure.stores.auth_store import AuthStore
-from lyra.infrastructure.stores.bot_agent_map import BotAgentMapStore
+from lyra.infrastructure.stores.base.bot_agent_map import BotAgentMapStore
+from lyra.infrastructure.stores.base.message_index import MessageIndex
+from lyra.infrastructure.stores.base.sqlite_base import SqliteStore
 from lyra.infrastructure.stores.bot_store import BotStore
 from lyra.infrastructure.stores.identity_alias_store import IdentityAliasStore
-from lyra.infrastructure.stores.message_index import MessageIndex
 from lyra.infrastructure.stores.message_index_kv import MessageIndexKvStore
 from lyra.infrastructure.stores.pairing import (
     PairingManager,
@@ -14,7 +15,6 @@ from lyra.infrastructure.stores.pairing import (
     set_pairing_manager,
 )
 from lyra.infrastructure.stores.prefs_store import PrefsStore, UserPrefs
-from lyra.infrastructure.stores.sqlite_base import SqliteStore
 from lyra.infrastructure.stores.thread_store import ThreadStore
 from lyra.infrastructure.stores.turn_store import TurnStore
