+++ src/lyra/infrastructure/stores/auth_store.py
@@ -8,7 +8,7 @@ from datetime import datetime, timezone
 from pathlib import Path
 
 from lyra.core.auth.trust import TrustLevel
-from lyra.infrastructure.stores.sqlite_base import SqliteStore
+from lyra.infrastructure.stores.base.sqlite_base import SqliteStore
 
 log = logging.getLogger(__name__)
