+++ src/lyra/infrastructure/stores/base/__init__.py
@@ -1,0 +1,2 @@
+"""
+"""
