+++ src/lyra/infrastructure/stores/base/sqlite_base.py
