+++ src/lyra/infrastructure/stores/turn_store.py
@@ -17,7 +17,7 @@ from datetime import UTC, datetime
 from typing import Literal
 
 from lyra.core.stores.turn_store_protocol import TurnRow
-from lyra.infrastructure.stores.sqlite_base import SqliteStore
+from lyra.infrastructure.stores.base.sqlite_base import SqliteStore
 from lyra.infrastructure.stores.turn_store_session import TurnStoreSessionMixin
 
 from .turn_store_queries import backfill_sessions, get_turns
