+++ src/lyra/nats/audio/__init__.py
@@ -1,0 +1,2 @@
+"""
+"""
