+++ src/lyra/nats/image/__init__.py
@@ -1,0 +1,2 @@
+"""
+"""
