+++ src/lyra/nats/image/nats_image_codec.py
