+++ src/lyra/nats/stt/__init__.py
@@ -1,0 +1,2 @@
+"""
+"""
