+++ src/lyra/nats/stt/nats_stt_codec.py
