+++ src/lyra/nats/stt/stt_helpers.py
