+++ tests/core/test_message_index.py
@@ -6,7 +6,7 @@ from datetime import UTC, datetime, timedelta
 
 import pytest
 
-from lyra.infrastructure.stores.message_index import MessageIndex
+from lyra.infrastructure.stores.base.message_index import MessageIndex
 
 
 @pytest.fixture
